ims-backend/requirements.txt
fastapi==0.110.2
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic-settings
psycopg2-binary
